Send jwt token with google oauth login
