--- backend/Model/passport.js
@@ -1,28 +1,0 @@
-import dotenv from 'dotenv';
-dotenv.config();
-import passport from 'passport';
-import { Strategy as GoogleStrategy } from 'passport-google-oauth20';
-
-passport.serializeUser(function (user, done) {
-  done(null, user);
-});
-
-passport.deserializeUser(function (user, done) {
-  done(null, user);
-});
-
-passport.use(
-  new GoogleStrategy(
-    {
-      clientID: process.env.GOOGLE_CLIENT_ID,
-      clientSecret: process.env.GOOGLE_CLIENT_SECRET,
-      callbackURL:
-        'http://localhost:5000/api/user/google/callback' ||
-        'https://dsa-competition-app.herokuapp.com/api/user/google/callback',
-      passReqToCallback: true,
-    },
-    function (request, accessToken, refreshToken, profile, done) {
-      return done(null, profile);
-    }
-  )
-);
