+++ backend/package.json
@@ -22,6 +22,7 @@
     "jsonwebtoken": "^8.5.1",
     "juice": "^8.0.0",
     "mongoose": "^6.2.9",
+    "morgan": "^1.10.0",
     "nodemailer": "^6.7.3",
     "nodemon": "^2.0.15",
     "passport": "^0.5.2",
