+++ backend/server.js
@@ -7,15 +7,19 @@ import yaml from 'yamljs';
 const swaggerJsDocs = yaml.load('./api.yaml');
 import cors from 'cors';
 import cookieSession from 'cookie-session';
-import './Model/passport.js';
+import morgan from 'morgan';
+
+import './config/googleAuthConfig.js';
 import './config/db.js';
 import './config/githubAuthConfig.js';
-import oauthRoutes from './Routes/oauthRoutes.js';
+
+import oauthRoutes from './Routes/googleAuthRoutes.js';
 import githubAuthRoutes from './Routes/githubAuthRoutes.js';
 import userRoutes from './Routes/userRoutes.js';
 import codeRoutes from './Routes/codeRoutes.js';
 
 const app = express();
+app.use(morgan('tiny'));
 app.use(cookieSession({ name: 'auth-session', keys: ['key1', 'key2'] }));
 app.use(passport.initialize());
 app.use(passport.session());
